document_service/requirements.txt
# Python文档处理微服务依赖
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2

# 文档处理
unstructured==0.11.2
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
python-magic==0.4.27

# OSS存储
oss2==2.18.1

# LLM集成
openai==1.3.5

# 数据库
aiomysql==0.2.0
aiosqlite==0.19.0

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
